added eslint config and clean eslint errors
