+++ package.json
@@ -21,12 +21,11 @@
     "@types/node": "^16.11.24",
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
-    "react-scripts": "5.0.0",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
     "eslint": "^7.11.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-airbnb-typescript": "^16.0.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-import-helpers": "^1.2.0",
@@ -35,6 +34,7 @@
     "eslint-plugin-react": "^7.27.1",
     "eslint-plugin-react-hooks": "^4.3.0",
     "prettier": "^2.4.1",
+    "react-scripts": "5.0.0",
     "typescript": "^4.5.5"
   },
   "scripts": {
@@ -44,7 +44,8 @@
     "eject": "react-scripts eject",
     "prettier": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,md,css,scss}\"",
     "prettier:check": "prettier --list-different \"src/**/*.{js,jsx,ts,tsx,json,md,css,scss}\"",
-    "lint" : "eslint ./src/**/*.{tsx,ts}"
+    "lint": "eslint ./src/**/*.{tsx,ts}",
+    "lint:fix": "eslint ./src/**/*.{tsx,ts} --fix"
   },
   "eslintConfig": {
     "extends": [
